+++ particlegan/__init__.py
@@ -8,6 +8,7 @@ from .diffusion import DDGAN
 from .discriminators import BatchDistanceDiscriminator, LinearSkipDiscriminator
 from .gan_loss import GANLoss
 from .grad_regularizers import GradientPenalty
+from .k3p import CriticAnchor, CriticSpikeGuard, DirectParticleResponse, LatentRowDamping
 from .locked_shared import LOCKED_SHARED, locked_adv_defaults, make_b_cap, make_gan_loss
 from .particle_prior import GaussianPrior, MoGParticlePrior, ParticlePrior, calibrate_mog_sigma
 from .recipes import Recipe, get_recipe, learning_rate_scale
@@ -17,6 +18,7 @@ from .vicreg_loss import ParticleRegularizer
 __all__ = [
     "ParticleEncoding", "particle_ae", "particle_vae",
     "ParticlePrior", "MoGParticlePrior", "calibrate_mog_sigma", "GaussianPrior", "GANLoss", "GradientPenalty",
+    "CriticAnchor", "CriticSpikeGuard", "LatentRowDamping", "DirectParticleResponse",
     "ParticleRegularizer", "DDGAN", "UCD", "ucd_labels", "ucd_loss", "ucd_scores",
     "Recipe", "get_recipe", "learning_rate_scale", "GANTrainer",
     "BatchDistanceDiscriminator", "LinearSkipDiscriminator",